+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.3.0
 - Added StarRail#fetchUser.
+- Use adm-zip library instead of unzipper.
 # 0.2.0
 - Added CharacterData#stars and LightConeData#stars.
 - Added LightConeData#superimpositions.